requirements.txt
aiosqlite==0.20.0
APScheduler==3.10.4
greenlet==3.0.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyaes==1.6.1
pymediainfo-pyrofork==6.0.1
pyrofork==2.3.19.post2
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
TgCrypto==1.2.5
typing_extensions==4.9.0
tzlocal==5.2
uvloop==0.19.0
